+++ extras/tests/JsonDocument/assignment.cpp
@@ -9,63 +9,59 @@
 
 using ArduinoJson::detail::sizeofArray;
 using ArduinoJson::detail::sizeofObject;
+using ArduinoJson::detail::sizeofString;
 
 TEST_CASE("JsonDocument assignment") {
   SpyingAllocator spyingAllocator;
 
   SECTION("Copy assignment same capacity") {
-    {
-      JsonDocument doc1(1024, &spyingAllocator);
-      deserializeJson(doc1, "{\"hello\":\"world\"}");
-      JsonDocument doc2(1024, &spyingAllocator);
+    JsonDocument doc1(1024, &spyingAllocator);
+    deserializeJson(doc1, "{\"hello\":\"world\"}");
+    JsonDocument doc2(1024, &spyingAllocator);
+    spyingAllocator.clearLog();
 
-      doc2 = doc1;
+    doc2 = doc1;
 
-      REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
-    }
-    REQUIRE(spyingAllocator.log() == AllocatorLog()
-                                         << AllocatorLog::Allocate(1024)
-                                         << AllocatorLog::Allocate(1024)
-                                         << AllocatorLog::Deallocate(1024)
-                                         << AllocatorLog::Deallocate(1024));
+    REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
+
+    REQUIRE(spyingAllocator.log() ==
+            AllocatorLog() << AllocatorLog::Allocate(sizeofString(5))  // hello
+                           << AllocatorLog::Allocate(sizeofString(5))  // world
+    );
   }
 
   SECTION("Copy assignment reallocates when capacity is smaller") {
-    {
-      JsonDocument doc1(4096, &spyingAllocator);
-      deserializeJson(doc1, "{\"hello\":\"world\"}");
-      JsonDocument doc2(8, &spyingAllocator);
+    JsonDocument doc1(4096, &spyingAllocator);
+    deserializeJson(doc1, "{\"hello\":\"world\"}");
+    JsonDocument doc2(8, &spyingAllocator);
+    spyingAllocator.clearLog();
 
-      doc2 = doc1;
+    doc2 = doc1;
 
-      REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
-    }
-    REQUIRE(spyingAllocator.log() == AllocatorLog()
-                                         << AllocatorLog::Allocate(4096)
-                                         << AllocatorLog::Allocate(8)
-                                         << AllocatorLog::Deallocate(8)
-                                         << AllocatorLog::Allocate(4096)
-                                         << AllocatorLog::Deallocate(4096)
-                                         << AllocatorLog::Deallocate(4096));
+    REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
+    REQUIRE(spyingAllocator.log() ==
+            AllocatorLog() << AllocatorLog::Deallocate(8)
+                           << AllocatorLog::Allocate(4096)
+                           << AllocatorLog::Allocate(sizeofString(5))  // hello
+                           << AllocatorLog::Allocate(sizeofString(5))  // world
+    );
   }
 
   SECTION("Copy assignment reallocates when capacity is larger") {
-    {
-      JsonDocument doc1(1024, &spyingAllocator);
-      deserializeJson(doc1, "{\"hello\":\"world\"}");
-      JsonDocument doc2(4096, &spyingAllocator);
+    JsonDocument doc1(1024, &spyingAllocator);
+    deserializeJson(doc1, "{\"hello\":\"world\"}");
+    JsonDocument doc2(4096, &spyingAllocator);
+    spyingAllocator.clearLog();
 
-      doc2 = doc1;
+    doc2 = doc1;
 
-      REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
-    }
-    REQUIRE(spyingAllocator.log() == AllocatorLog()
-                                         << AllocatorLog::Allocate(1024)
-                                         << AllocatorLog::Allocate(4096)
-                                         << AllocatorLog::Deallocate(4096)
-                                         << AllocatorLog::Allocate(1024)
-                                         << AllocatorLog::Deallocate(1024)
-                                         << AllocatorLog::Deallocate(1024));
+    REQUIRE(doc2.as<std::string>() == "{\"hello\":\"world\"}");
+    REQUIRE(spyingAllocator.log() ==
+            AllocatorLog() << AllocatorLog::Deallocate(4096)
+                           << AllocatorLog::Allocate(1024)
+                           << AllocatorLog::Allocate(sizeofString(5))  // hello
+                           << AllocatorLog::Allocate(sizeofString(5))  // world
+    );
   }
 
   SECTION("Move assign") {
@@ -79,11 +75,13 @@ TEST_CASE("JsonDocument assignment") {
       REQUIRE(doc2.as<std::string>() == "The size of this string is 32!!");
       REQUIRE(doc1.as<std::string>() == "null");
     }
-    REQUIRE(spyingAllocator.log() == AllocatorLog()
-                                         << AllocatorLog::Allocate(4096)
-                                         << AllocatorLog::Allocate(8)
-                                         << AllocatorLog::Deallocate(8)
-                                         << AllocatorLog::Deallocate(4096));
+    REQUIRE(spyingAllocator.log() ==
+            AllocatorLog() << AllocatorLog::Allocate(4096)
+                           << AllocatorLog::Allocate(sizeofString(31))
+                           << AllocatorLog::Allocate(8)
+                           << AllocatorLog::Deallocate(8)
+                           << AllocatorLog::Deallocate(sizeofString(31))
+                           << AllocatorLog::Deallocate(4096));
   }
 
   SECTION("Assign from JsonObject") {