+++ src/ArduinoJson/StringStorage/StringCopier.hpp
@@ -10,20 +10,31 @@ ARDUINOJSON_BEGIN_PRIVATE_NAMESPACE
 
 class StringCopier {
  public:
+  static const size_t initialCapacity = 31;
+
   StringCopier(MemoryPool* pool) : _pool(pool) {}
 
+  ~StringCopier() {
+    if (_node)
+      _pool->deallocString(_node);
+  }
+
   void startString() {
-    _pool->getFreeZone(&_ptr, &_capacity);
     _size = 0;
-    if (_capacity == 0)
-      _pool->markAsOverflowed();
+    if (!_node)
+      _node = _pool->allocString(initialCapacity);
   }
 
   JsonString save() {
-    ARDUINOJSON_ASSERT(_ptr);
-    ARDUINOJSON_ASSERT(_size < _capacity);  // needs room for the terminator
-    return JsonString(_pool->saveStringFromFreeZone(_size), _size,
-                      JsonString::Copied);
+    ARDUINOJSON_ASSERT(_node != nullptr);
+    _node->data[_size] = 0;
+    StringNode* node = _pool->findString(adaptString(_node->data, _size));
+    if (!node) {
+      node = _pool->reallocString(_node, _size);
+      _pool->addStringToList(node);
+      _node = nullptr;  // next time we need a new string
+    }
+    return JsonString(node->data, node->length, JsonString::Copied);
   }
 
   void append(const char* s) {
@@ -32,19 +43,19 @@ class StringCopier {
   }
 
   void append(const char* s, size_t n) {
-    while (n-- > 0)
+    while (n-- > 0)  // TODO: memcpy
       append(*s++);
   }
 
   void append(char c) {
-    if (_size + 1 < _capacity)
-      _ptr[_size++] = c;
-    else
-      _pool->markAsOverflowed();
+    if (_node && _size == _node->length)
+      _node = _pool->reallocString(_node, _size * 2U + 1);
+    if (_node)
+      _node->data[_size++] = c;
   }
 
   bool isValid() const {
-    return !_pool->overflowed();
+    return _node != nullptr;
   }
 
   size_t size() const {
@@ -52,21 +63,15 @@ class StringCopier {
   }
 
   JsonString str() const {
-    ARDUINOJSON_ASSERT(_ptr);
-    ARDUINOJSON_ASSERT(_size < _capacity);
-    _ptr[_size] = 0;
-    return JsonString(_ptr, _size, JsonString::Copied);
+    ARDUINOJSON_ASSERT(_node != nullptr);
+    _node->data[_size] = 0;
+    return JsonString(_node->data, _size, JsonString::Copied);
   }
 
  private:
   MemoryPool* _pool;
-
-  // These fields aren't initialized by the constructor but startString()
-  //
-  // NOLINTNEXTLINE(clang-analyzer-optin.cplusplus.UninitializedObject)
-  char* _ptr;
-  // NOLINTNEXTLINE(clang-analyzer-optin.cplusplus.UninitializedObject)
-  size_t _size, _capacity;
+  StringNode* _node = nullptr;
+  size_t _size = 0;
 };
 
 ARDUINOJSON_END_PRIVATE_NAMESPACE